refactoring: due to the module name changes
